+++ src/examples/kotlin/ru/yandex/money/tools/grafana/dsl/examples/GrafanaDemoLayouts.kts
@@ -7,9 +7,12 @@ import ru.yandex.money.tools.grafana.dsl.json.set
 import ru.yandex.money.tools.grafana.dsl.metrics.functions.StringMetric
 import ru.yandex.money.tools.grafana.dsl.metrics.functions.alias
 import ru.yandex.money.tools.grafana.dsl.metrics.functions.aliasByNode
+import ru.yandex.money.tools.grafana.dsl.metrics.functions.asPercent
 import ru.yandex.money.tools.grafana.dsl.metrics.functions.averageSeries
 import ru.yandex.money.tools.grafana.dsl.metrics.functions.groupByNodes
 import ru.yandex.money.tools.grafana.dsl.metrics.functions.movingMedian
+import ru.yandex.money.tools.grafana.dsl.metrics.functions.perSecond
+import ru.yandex.money.tools.grafana.dsl.metrics.functions.sortByTotal
 import ru.yandex.money.tools.grafana.dsl.panels.Color
 import ru.yandex.money.tools.grafana.dsl.panels.Legend
 import ru.yandex.money.tools.grafana.dsl.panels.NullPointMode
@@ -42,9 +45,11 @@ dashboard(title = "Grafana Demo Layouts") {
     }
 
     panels {
-        row(title = "Single Stat Metrics") { // Create row with single-stat metrics
+        row(title = "Single Stat Metrics") {
+            // Create row with single-stat metrics
 
-            val singleStat = { title: String, metric: String -> // Create fabric-function for single-stat panels
+            val singleStat = { title: String, metric: String ->
+                // Create fabric-function for single-stat panels
 
                 metricPanel(title = title) {
 
@@ -54,7 +59,8 @@ dashboard(title = "Grafana Demo Layouts") {
                      */
                     bounds = 6 to 3
 
-                    properties { // Override some properties in JSON
+                    properties {
+                        // Override some properties in JSON
                         it["type"] = "singlestat"
                         it["nullPointMode"] = "connected"
                         it["valueName"] = "avg"
@@ -68,7 +74,8 @@ dashboard(title = "Grafana Demo Layouts") {
                     }
 
                     metrics {
-                        metric("A") { // Define a metric referencing above one (referenceId must be uniq for panel)
+                        metric("A") {
+                            // Define a metric referencing above one (referenceId must be uniq for panel)
                             StringMetric(metric)
                         }
                     }
@@ -93,11 +100,13 @@ dashboard(title = "Grafana Demo Layouts") {
                 stacked = true // Metrics must not be stacked onto Ox axis, and should overlap each other
                 legend = Legend.EMPTY // Show only aliases for metrics
                 fill = 1 // Fill rate
+                staircase = true // enable staircase line vizualization
+                decimals = null // set automatic decimal precision
                 nullPointMode = NullPointMode.NULL // How to show null values
                 points = true // Show chart points
                 pointradius = 2 // Chart points radius
-                leftYAxis = YAxis(format = YAxis.MILLISECONDS, min = 0, max = 100)
-                rightYAxis = YAxis()
+                leftYAxis = YAxis(format = YAxis.MILLISECONDS, min = 0, max = 100, scale = YAxis.Scale.LOG10)
+                rightYAxis = YAxis(format = YAxis.PERCENT, scale = YAxis.Scale.LOG1024)
 
                 aliasColors {
                     "some metric" to Color.GREEN // Use predefined color
@@ -105,24 +114,34 @@ dashboard(title = "Grafana Demo Layouts") {
                 }
 
                 metrics {
-                    metric("A") {
-                        /*
-                        use infix style function to define metric
-                         */
-                        "apps.backend.*.counters.requests.count" movingMedian medianInterval aliasByNode 2 alias "some metric"
-                    }
+                    /*
+                        use infix style function to define metric and save in variable
+                     */
+                    val refMetric = "apps.backend.*.counters.requests.count" movingMedian medianInterval aliasByNode 2 alias "some metric"
 
                     metric("B") {
                         "*.another.metric.mean"
-                                .groupByNodes(0)
-                                .averageSeries() // show average value for metric
-                                .alias("another metric") // define alias
+                            .groupByNodes(0)
+                            .averageSeries() // show average value for metric
+                            .alias("another metric") // define alias
+                            .perSecond() // show metric on value per second
+                            .sortByTotal() // sort metrics in descending order by the sum of values across time period
+                    }
+
+                    metric("C") {
+                        "apps.backend.*.counters.login" alias "Some" override { bars = true } // override parameter 'bars'
+                    }
+
+                    metric("D") {
+                        // represents the ratio of metric to metric variable 'refMetric' as a percentage
+                        "apps.backend.*.counters.responses.count" aliasByNode 1 asPercent refMetric
                     }
                 }
             }
         }
 
-        row(title = "Repeated row", repeatFor = hosts) { // That row with all nested panels will be repeated for each values of hosts
+        row(title = "Repeated row", repeatFor = hosts) {
+            // That row with all nested panels will be repeated for each values of hosts
 
             graphPanel(title = "Graph panel with per-host metrics") {