+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/panels/GraphPanelBuilder.kt
@@ -6,6 +6,10 @@ import ru.yandex.money.tools.grafana.dsl.datasource.Graphite
 import ru.yandex.money.tools.grafana.dsl.metrics.Metrics
 import ru.yandex.money.tools.grafana.dsl.metrics.MetricsBuilder
 import ru.yandex.money.tools.grafana.dsl.metrics.ReferenceMetricsHolder
+import ru.yandex.money.tools.grafana.dsl.metrics.functions.Alias
+import ru.yandex.money.tools.grafana.dsl.panels.graph.display.drawoptions.HoverTooltip
+import ru.yandex.money.tools.grafana.dsl.panels.graph.display.seriesoverrides.SeriesOverride
+import ru.yandex.money.tools.grafana.dsl.panels.graph.display.seriesoverrides.SeriesOverrideBuilder
 import ru.yandex.money.tools.grafana.dsl.time.Duration
 
 class GraphPanelBuilder(private val title: String) : PanelBuilder {
@@ -26,6 +30,8 @@ class GraphPanelBuilder(private val title: String) : PanelBuilder {
 
     var legend = Legend.DEFAULT
 
+    var decimals: Int? = 2
+
     var points = false
 
     var pointradius = 5
@@ -34,11 +40,19 @@ class GraphPanelBuilder(private val title: String) : PanelBuilder {
 
     var fill = 0
 
+    var lineWidth = 2
+
+    var staircase = false
+
+    var hoverTooltip = HoverTooltip()
+
     var aliasColors: AliasColors? = null
 
     var leftYAxis: YAxis? = null
 
     var rightYAxis: YAxis? = null
+
+    private val seriesOverrides: MutableList<SeriesOverride> = mutableListOf()
 
     override fun properties(propertiesSetter: (JSONObject) -> Unit) {
         propertiesSetters += propertiesSetter
@@ -56,6 +70,13 @@ class GraphPanelBuilder(private val title: String) : PanelBuilder {
         metrics += builder.metrics
     }
 
+    infix fun Alias.override(build: SeriesOverrideBuilder.() -> Unit): Alias {
+        val builder = SeriesOverrideBuilder(this.aliasName)
+        builder.build()
+        seriesOverrides += builder.createSeriesOverride()
+        return this
+    }
+
     internal fun createPanel() = AdditionalPropertiesPanel(
             GraphPanel(
                     MetricPanel(
@@ -71,13 +92,18 @@ class GraphPanelBuilder(private val title: String) : PanelBuilder {
                     timeShift = timeShift,
                     stack = stacked,
                     legend = legend,
+                    decimals = decimals,
                     points = points,
                     pointradius = pointradius,
                     nullPointMode = nullPointMode,
                     fill = fill,
+                    lineWidth = lineWidth,
+                    staircase = staircase,
+                    hoverTooltip = hoverTooltip,
                     aliasColors = aliasColors,
                     leftYAxis = leftYAxis,
-                    rightYAxis = rightYAxis
+                    rightYAxis = rightYAxis,
+                    seriesOverrides = seriesOverrides
             )
     ) { json -> propertiesSetters.forEach { it(json) } }
 }